requirement.txt
# Core RAG and LLM libraries
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.2.0
tiktoken>=0.5.0

# Embeddings and vector store
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain-huggingface>=0.0.3

# Transcript fetching and video processing
youtube-transcript-api>=0.6.1
pytube>=15.0.0
requests>=2.31.0

# Web framework and API
flask>=3.0.0
flask-cors>=4.0.0

# Text processing and chunking
nltk>=3.8.1
spacy>=3.7.0

# Environment and configuration
python-dotenv>=1.0.1

# Data processing and utilities
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0

# Logging and error handling
loguru>=0.7.2

# JSON and data handling
json5>=0.9.14

# HTTP client for API calls
httpx>=0.25.0

# Optional: For better async support
asyncio>=3.4.3

# Optional: For PDF processing (if needed later)
PyPDF2>=3.0.1

# Optional: For advanced text splitting
unstructured>=0.10.0

# Optional: For better tokenization
transformers>=4.30.0

# Optional: For caching
diskcache>=5.6.3
